refactor: separated raster layer edit components as a RasterLegendEdit
